fastapi/requirements.txt
fastapi
uvicorn
transformers
pdfminer.six
torch
python-multipart